feat: MemberDetail props 적용
