+++ frontend/src/components/admin/dummy/dummy.tsx
@@ -1,0 +1,30 @@
+const UserData = {
+  content: [
+    {
+      grade: 'PRESIDENT',
+      name: '윤찬호',
+      id: 'C011001',
+      major: '컴퓨터공학과',
+      phone_number: '010-1111-1111',
+      acceptDate: '2024-01-08 16:49',
+    },
+    {
+      grade: 'GRADUATE',
+      name: '김진호',
+      id: 'B731000',
+      major: '컴퓨터공학과',
+      phone_number: '010-1111-1111',
+      acceptDate: '2024-01-08 16:49',
+    },
+    {
+      grade: 'GRADUATE',
+      name: '최세호',
+      id: 'B811197',
+      major: '컴퓨터공학과',
+      phone_number: '010-1111-1111',
+      acceptDate: '2024-01-08 16:49',
+    },
+  ],
+};
+
+export default UserData;
